requirements.txt
openai>=2.8.1
python-dotenv>=1.2.1
requests>=2.32.5
langchain>=1.1.0
langchain-core>=1.1.0
requests>=2.32.5
langchain-openai>=1.1.0
